Brings in #327 (PyVista), #334 (GNU compiler warnings, filename truncation) and
the ICL module-version update.

Seven conflicts, all where master fixed warnings in code this branch retires or
renames. Resolved by keeping this branch's structure and adopting master's fixes
wherever they still apply:

- modinlet.f90: deletion kept. Master only fixed warnings in it; Phase 2 removes
  the file because the generator is unreachable (iinletgen is in no namelist).
- modibm.f90: `subroutine bottom` stays deleted (Phase 1); tau_x/tau_y/tau_z/
  thl_flux stay retired. Took master's drop of MPI_SUM, which survives here only
  in commented-out MPI_ALLREDUCE calls.
- inflow.f90 vs master's moddriver.f90: kept the rename (initinflow), took
  master's drop of the unused `pi`.
- modsubgrid.f90: kept thvf from modfields (Task 4; modsurfdata is gone), took
  master's drop of numoli and mpi_sum -- numoli survives only in two commented
  lines, mpi_sum only on the use line.
- modstatsdump.f90: kept modbasestate/thv_b over master's modsurfdata/thls, took
  the mpi_sum drop.
- modsave.f90: master's side could not compile here -- iinletgen and nstore no
  longer exist in modglobal. None of iinletgen/nstore/numol/grav is referenced in
  modsave and timee was already imported on the line above, so the whole
  continuation goes, which is what #334 was doing anyway.
- modstartup.f90: BCbot* stay retired; the inlet-generator locals (Utavin,
  Urecin, ...) and name4 stay deleted. Kept master's filename-truncation fix
  (#333): startfile/author broadcast with len() instead of hardcoded 50/80, and
  `character(len(startfile)) :: name` in readrestartfiles, where name = startfile.
  In readinitfiles grav/pref0 (ours) and Uinf (master's) are all unused -- they
  appear only in commented-out lines; Uinf was live for master because its
  inlet-generator block used it, and that block is retired here.

Verified on CX3 (Intel 2021a, Debug): builds clean, basestate unit test and
retired-keys contract both pass.

+++ src/initfac.f90
@@ -56,8 +56,10 @@
       real, allocatable    :: netsw(:) !net shortwave radiation on facets
       real, allocatable    :: facLWin(:) !incoming longwave on facets [W/m2]
       real, allocatable    :: vfsparse(:)
-      real, allocatable    :: ivfsparse(:)
-      real, allocatable    :: jvfsparse(:)
+      ! facet indices of the sparse view-factor entries: integer, matching the
+      ! `%d %d %.6f` layout written by the pre-processing.
+      integer, allocatable :: ivfsparse(:)
+      integer, allocatable :: jvfsparse(:)
       !temperature
       real, allocatable    :: Tfacinit(:) !initial facet temperatures
       real, allocatable    :: Tfacinit_layers(:,:) !initial facet temperatures
